+++ pages/details/[id].js
@@ -28,7 +28,7 @@ export default function DetailsPage() {
   return (
     <StyledDetailsPage>
       <StyledDetailsHeader>
-        <StyledBackLink href={"/"} onClick={() => router.back()}>
+        <StyledBackLink href={"#"} onClick={() => router.back()}>
           <TiArrowBackOutline />
           Back
         </StyledBackLink>
@@ -163,7 +163,7 @@ const StyledDetailsHeader = styled.header`
   }
 `;
 
-const StyledBackLink = styled(Link)`
+const StyledBackLink = styled.a`
   text-decoration: none;
   background-color: #fff7f7;
   padding-left: 20px;
